Convert to app router (#39)
* Move homepage to app

* Move dashboard to app

Adds LoginLinks component

* Update to nextjs 14.0.3

* switch to useParams

* Colocate components

* Move to (app) grouped folder

* Move to (auth) grouped folder

create not-found

* Fix import

* Use existing export syntax

* Fix syntax error

* Fix error with loading Nunito font

* Drop @header layout file

* Move header layout back to main layout

* Update global.css

* Colocate DashboardLayout

* Add @header back the correct way

* remove unnecessary DashboardLayout

* Add loading to avoid page flicker

* wip

---------
