+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@headlessui/react": "^1.4.2",
     "axios": "^0.21.1",
-    "next": "^13.0.3",
+    "next": "^14.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "swr": "^1.3.0"
@@ -25,7 +25,7 @@
     "autoprefixer": "^10.4.2",
     "babel-eslint": "^10.1.0",
     "eslint": "^8.27.0",
-    "eslint-config-next": "13.0.3",
+    "eslint-config-next": "^14.0.3",
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-prettier": "^3.3.0",
     "postcss": "^8.4.6",
