+++ src/app/(app)/@header/dashboard/page.js
@@ -1,0 +1,9 @@
+const Header = () => {
+    return (
+        <h2 className="font-semibold text-xl text-gray-800 leading-tight">
+            Dashboard
+        </h2>
+    )
+}
+
+export default Header
