+++ src/app/(app)/@header/default.js
@@ -1,0 +1,9 @@
+const Header = () => {
+    return (
+        <h2 className="font-semibold text-xl text-gray-800 leading-tight">
+            Dashboard
+        </h2>
+    )
+}
+
+export default Header
