+++ src/app/(app)/layout.js
@@ -1,21 +1,26 @@
-import Navigation from '@/components/Layouts/Navigation'
-import { useAuth } from '@/hooks/auth'
+'use client'
 
-const AppLayout = ({ header, children }) => {
+import { useAuth } from '@/hooks/auth'
+import Navigation from '@/app/(app)/Navigation'
+import Loading from '@/app/(app)/Loading'
+
+const AppLayout = ({ children, header }) => {
     const { user } = useAuth({ middleware: 'auth' })
+
+    if (!user) {
+        return <Loading />
+    }
 
     return (
         <div className="min-h-screen bg-gray-100">
             <Navigation user={user} />
 
-            {/* Page Heading */}
             <header className="bg-white shadow">
                 <div className="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8">
                     {header}
                 </div>
             </header>
 
-            {/* Page Content */}
             <main>{children}</main>
         </div>
     )
