+++ src/app/global.css
@@ -1,0 +1,5 @@
+@import url("https://fonts.bunny.net/css2?family=Nunito:wght@400;600;700&display=swap");
+
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
