+++ src/app/layout.js
@@ -1,0 +1,14 @@
+import '@/app/global.css'
+
+export const metadata = {
+    title: 'Laravel',
+}
+const RootLayout = ({ children }) => {
+    return (
+        <html lang="en">
+            <body className="antialiased">{children}</body>
+        </html>
+    )
+}
+
+export default RootLayout
