+++ src/app/page.js
@@ -1,40 +1,14 @@
-import Head from 'next/head'
-import Link from 'next/link'
-import { useAuth } from '@/hooks/auth'
+import LoginLinks from '@/app/LoginLinks'
 
-export default function Home() {
-    const { user } = useAuth({ middleware: 'guest' })
+export const metadata = {
+    title: 'Laravel',
+}
 
+const Home = () => {
     return (
         <>
-            <Head>
-                <title>Laravel</title>
-            </Head>
-
             <div className="relative flex items-top justify-center min-h-screen bg-gray-100 dark:bg-gray-900 sm:items-center sm:pt-0">
-                <div className="hidden fixed top-0 right-0 px-6 py-4 sm:block">
-                    {user ? (
-                        <Link
-                            href="/dashboard"
-                            className="ml-4 text-sm text-gray-700 underline">
-                            Dashboard
-                        </Link>
-                    ) : (
-                        <>
-                            <Link
-                                href="/login"
-                                className="text-sm text-gray-700 underline">
-                                Login
-                            </Link>
-
-                            <Link
-                                href="/register"
-                                className="ml-4 text-sm text-gray-700 underline">
-                                Register
-                            </Link>
-                        </>
-                    )}
-                </div>
+                <LoginLinks />
 
                 <div className="max-w-6xl mx-auto sm:px-6 lg:px-8">
                     <div className="flex justify-center pt-8 sm:justify-start sm:pt-0">
@@ -291,3 +265,5 @@ export default function Home() {
         </>
     )
 }
+
+export default Home
