--- src/components/Layouts/GuestLayout.js
@@ -1,17 +1,0 @@
-import Head from 'next/head'
-
-const GuestLayout = ({ children }) => {
-    return (
-        <div>
-            <Head>
-                <title>Laravel</title>
-            </Head>
-
-            <div className="font-sans text-gray-900 antialiased">
-                {children}
-            </div>
-        </div>
-    )
-}
-
-export default GuestLayout
