--- src/pages/_app.js
@@ -1,5 +1,0 @@
-import 'tailwindcss/tailwind.css'
-
-const App = ({ Component, pageProps }) => <Component {...pageProps} />
-
-export default App
