--- src/pages/_document.js
@@ -1,27 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document'
-
-class MyDocument extends Document {
-    static async getInitialProps(ctx) {
-        const initialProps = await Document.getInitialProps(ctx)
-        return { ...initialProps }
-    }
-
-    render() {
-        return (
-            <Html>
-                <Head>
-                    <link
-                        href="https://fonts.bunny.net/css2?family=Nunito:wght@400;600;700&display=swap"
-                        rel="stylesheet"
-                    />
-                </Head>
-                <body className="antialiased">
-                    <Main />
-                    <NextScript />
-                </body>
-            </Html>
-        )
-    }
-}
-
-export default MyDocument
